requirements.txt
Django==4.1.5
psycopg2==2.9.5
djangorestframework==3.14.0
python-decouple==3.7
djangorestframework_simplejwt